Enity Viewed

+++ src/main/java/com/capgemini/film_rental/entity/Payment.java
@@ -10,7 +10,7 @@ public class Payment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "payment_id")
-    private Long paymentId;
+    private int paymentId;
 
     @ManyToOne
     @JoinColumn(name = "customer_id", nullable = false,
@@ -41,8 +41,8 @@ public class Payment {
     public Payment() {}
 
     // Getters and Setters
-    public Long getPaymentId() { return paymentId; }
-    public void setPaymentId(Long paymentId) { this.paymentId = paymentId; }
+    public int getPaymentId() { return paymentId; }
+    public void setPaymentId(int paymentId) { this.paymentId = paymentId; }
 
     public Customer getCustomer() { return customer; }
     public void setCustomer(Customer customer) { this.customer = customer; }
